+++ cutils/src/main/java/com/carlos/cutils/base/activity/CBaseEditActivity.kt
@@ -1,4 +1,4 @@
-package com.carlos.cutils.base
+package com.carlos.cutils.base.activity
 
 import android.app.Activity
 import android.content.Context