+++ cutils/src/main/java/com/carlos/cutils/extend/ViewExtend.kt
@@ -3,6 +3,7 @@ package com.carlos.cutils.extend
 import android.view.View
 import android.widget.Checkable
 import com.carlos.cutils.util.LogUtils
+import com.carlos.cutils.util.SoundPoolUtils
 
 /**
  * Github: https://github.com/xbdcc/.
@@ -17,6 +18,17 @@ inline fun <T : View> T.singleClick(spaceTime: Long = 100, crossinline block: (T
             block(this)
             LogUtils.d("点击了")
         }
+    }
+}
+
+inline fun <T : View> T.clickSound(resId: Int, isEnd: Boolean = true, crossinline block: (T) -> Unit) {
+    setOnClickListener {
+        if (!isEnd)
+            SoundPoolUtils.play(resId, it.context)
+        block(this)
+        if (isEnd)
+            SoundPoolUtils.play(resId, it.context)
+        LogUtils.d("点击后响声")
     }
 }